@@ -11,52 +11,53 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/core": "^0.7.5",
-    "@angular/animations": "^7.2.12",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.12",
-    "@angular/compiler": "^7.2.12",
-    "@angular/core": "^7.2.12",
-    "@angular/forms": "^7.2.12",
-    "@angular/http": "^7.2.12",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.12",
-    "@angular/platform-browser-dynamic": "^7.2.12",
-    "@angular/router": "^7.2.12",
-    "@moebius/ng-validators": "^1.1.1",
-    "@webcomponents/custom-elements": "^1.2.2",
-    "angular-oauth2-oidc": "^5.0.2",
-    "bootstrap": "^4.3.1",
-    "core-js": "^2.6.5",
+    "@angular-devkit/core": "^8.3.23",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
+    "@moebius/ng-validators": "^2.0.2",
+    "@webcomponents/custom-elements": "^1.3.2",
+    "angular-oauth2-oidc": "^8.0.4",
+    "bootstrap": "^4.4.1",
+    "core-js": "^3.6.4",
     "jwt-decode": "^2.2.0",
-    "ngx-monaco-editor": "^6.0.0",
-    "node-pre-gyp": "^0.10.3",
-    "npm-check-updates": "^2.15.0",
-    "rxjs": "^6.4.0",
-    "zone.js": "^0.8.29"
+    "moment": "^2.24.0",
+    "ngx-monaco-editor": "^8.1.1",
+    "node-pre-gyp": "^0.14.0",
+    "npm-check-updates": "^4.0.1",
+    "rxjs": "^6.5.4",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.8",
-    "@angular/cli": "^7.3.8",
-    "@angular/compiler-cli": "^7.2.12",
-    "@angular/language-service": "^7.2.12",
-    "@types/jasmine": "^2.8.16",
-    "@types/jasminewd2": "^2.0.6",
+    "@angular-devkit/build-angular": "^0.803.23",
+    "@angular/cli": "^8.3.23",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
+    "@types/jasmine": "^3.5.1",
+    "@types/jasminewd2": "^2.0.8",
     "@types/jwt-decode": "^2.2.1",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
-    "hoek": "^5.0.4",
-    "jasmine-core": "~2.99.1",
+    "@types/node": "~13.5.0",
+    "@typescript-eslint/eslint-plugin": "^2.17.0",
+    "@typescript-eslint/parser": "^2.17.0",
+    "eslint": "^6.8.0",
+    "eslint-plugin-import": "^2.20.0",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^2.0.5",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-swagger-gen": "^1.7.0",
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "^2.1.1",
+    "karma-jasmine": "~3.1.0",
+    "karma-jasmine-html-reporter": "^1.5.1",
+    "ng-swagger-gen": "^1.8.1",
     "protractor": "^5.4.2",
-    "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "^3.2.4"
+    "ts-node": "~8.6.2",
+    "typescript": "~3.5.0"
   }
 }
